chore: move legacy diff tests into misc sections
