+++ packages/sdk/src/diff/target.test.ts
@@ -1,9 +1,16 @@
 import { expect, it, suite } from "vitest"
-import { Clearance, ExecutionOptions, Target } from "zodiac-roles-deployments"
+import {
+  Clearance,
+  ExecutionOptions,
+  Operator,
+  ParameterType,
+  Target,
+} from "zodiac-roles-deployments"
 
 import { ZeroAddress, ZeroHash } from "ethers"
 
 import { diffTarget, diffTargets } from "./target"
+import { normalizeCondition } from "../conditions"
 
 const roleKey = ZeroHash
 const address = ZeroAddress as `0x${string}`
@@ -497,6 +504,229 @@ suite("diffTargets", () => {
           roleKey,
           targetAddress: "0x03",
           executionOptions: ExecutionOptions.None,
+        },
+      ],
+    })
+  })
+})
+
+suite("diffTarget - Misc", () => {
+  it("should revoke function-scoped targets", () => {
+    const before: Target[] = [
+      {
+        address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        clearance: 2,
+        executionOptions: 0,
+        functions: [
+          {
+            selector: "0x095ea7b3",
+            executionOptions: 0,
+            wildcarded: false,
+            condition: {
+              paramType: ParameterType.None,
+              operator: Operator.Pass,
+            },
+          },
+        ],
+      },
+    ]
+    const after: Target[] = []
+
+    const { minus, plus } = diffTargets({ roleKey, prev: before, next: after })
+
+    expect(minus).to.deep.equal([
+      {
+        call: "revokeTarget",
+        roleKey,
+        targetAddress: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+      },
+      {
+        call: "revokeFunction",
+        roleKey,
+        targetAddress: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        selector: "0x095ea7b3",
+      },
+    ])
+
+    expect(plus).to.deep.equal([])
+  })
+
+  it("should patch conditions", () => {
+    const before: Target[] = [
+      {
+        address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        clearance: 2,
+        executionOptions: 0,
+        functions: [
+          {
+            selector: "0x095ea7b3",
+            executionOptions: 0,
+            wildcarded: false,
+            condition: {
+              paramType: ParameterType.None,
+              operator: Operator.Pass,
+            },
+          },
+        ],
+      },
+    ]
+
+    const after: Target[] = [
+      {
+        address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        clearance: 2,
+        executionOptions: 0,
+        functions: [
+          {
+            selector: "0x095ea7b3",
+            executionOptions: 0,
+            wildcarded: false,
+            condition: {
+              paramType: ParameterType.None,
+              operator: Operator.Or,
+              children: [],
+            },
+          },
+        ],
+      },
+    ]
+
+    const { minus, plus } = diffTargets({ roleKey, prev: before, next: after })
+
+    expect(minus).to.deep.equal([])
+
+    expect(plus).to.deep.equal([
+      {
+        call: "scopeFunction",
+        roleKey,
+        targetAddress: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        selector: "0x095ea7b3",
+        executionOptions: 0,
+        condition: normalizeCondition({
+          paramType: ParameterType.None,
+          operator: Operator.Or,
+          children: [],
+        }),
+      },
+    ])
+  })
+
+  it("should correctly diff target-cleared targets", () => {
+    const a: Target[] = [
+      {
+        address: "0x1",
+        clearance: Clearance.Target,
+        executionOptions: ExecutionOptions.None,
+        functions: [],
+      },
+      {
+        address: "0x2",
+        clearance: Clearance.Target,
+        executionOptions: ExecutionOptions.None,
+        functions: [],
+      },
+    ]
+
+    const b: Target[] = [
+      {
+        address: "0x1",
+        clearance: Clearance.Target,
+        executionOptions: ExecutionOptions.None,
+        functions: [],
+      },
+      {
+        address: "0x2",
+        clearance: Clearance.Target,
+        executionOptions: ExecutionOptions.Both,
+        functions: [],
+      },
+    ]
+
+    expect(diffTargets({ roleKey, prev: a, next: b })).to.deep.equal({
+      minus: [],
+      plus: [
+        {
+          call: "allowTarget",
+          roleKey,
+          targetAddress: "0x2",
+          executionOptions: ExecutionOptions.Both,
+        },
+      ],
+    })
+
+    expect(diffTargets({ roleKey, prev: b, next: a })).to.deep.equal({
+      minus: [
+        {
+          call: "allowTarget",
+          roleKey,
+          targetAddress: "0x2",
+          executionOptions: ExecutionOptions.None,
+        },
+      ],
+      plus: [],
+    })
+  })
+
+  it("should diff functions based on all their properties including conditions", () => {
+    const a: Target[] = [
+      {
+        address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        clearance: 2,
+        executionOptions: 0,
+        functions: [
+          {
+            selector: "0x095ea7b3",
+            executionOptions: 0,
+            wildcarded: false,
+            condition: {
+              paramType: ParameterType.None,
+              operator: Operator.Pass,
+            },
+          },
+        ],
+      },
+    ]
+
+    const b: Target[] = [
+      {
+        address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+        clearance: 2,
+        executionOptions: 0,
+        functions: [
+          {
+            selector: "0x095ea7b3",
+            executionOptions: 0,
+            wildcarded: false,
+            condition: {
+              paramType: ParameterType.Dynamic,
+              operator: Operator.EqualTo,
+              compValue: "0x00",
+            },
+          },
+        ],
+      },
+    ]
+
+    //expect(diffTargets(a, a)).to.deep.equal([])
+    expect(diffTargets({ roleKey, prev: a, next: a })).to.deep.equal({
+      minus: [],
+      plus: [],
+    })
+
+    //expect(diffTargets(a, b)).to.deep.equal(a)
+    expect(diffTargets({ roleKey, prev: b, next: a })).to.deep.equal({
+      minus: [],
+      plus: [
+        {
+          call: "scopeFunction",
+          roleKey,
+          targetAddress: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb",
+          selector: "0x095ea7b3",
+          executionOptions: 0,
+          condition: normalizeCondition({
+            paramType: ParameterType.None,
+            operator: Operator.Pass,
+          }),
         },
       ],
     })
